+++ Sources/NablaMessagingUI/Scenes/ConversationList/Model/ConversationListItemViewModel.swift
@@ -3,7 +3,7 @@ import Foundation
 struct ConversationListItemViewModel {
     let avatar: AvatarViewModel
     let title: String
-    let lastMessage: String?
+    let subtitle: String?
     let lastUpdatedTime: String
     let isUnread: Bool
 }