+++ UITests/NablaUITestsHostApp/NablaUITestsHostApp/Mocks/Conversation+Mock.swift
@@ -7,7 +7,7 @@ extension Conversation {
         return Conversation(
             id: .init(),
             title: "Title \(conversationCount)",
-            description: "Description \(conversationCount)",
+            subtitle: "Subtitle \(conversationCount)",
             inboxPreviewTitle: "PreviewTitle \(conversationCount)",
             lastMessagePreview: "MessagePreview \(conversationCount)",
             lastModified: .init(timeIntervalSinceReferenceDate: 0),